Add scripts/Images; rename scripts/ArknightsGameData to GameData
